Text/requirements.txt
blas=1.0
ca-certificates=2018.03.07
certifi=2018.8.13
cffi=1.11.5
cudatoolkit=9.0
cycler=0.10.0
dbus=1.13.2
expat=2.2.5
fontconfig=2.13.0
freetype=2.9.1
glib=2.56.1
gst-plugins-base=1.14.0
gstreamer=1.14.0
icu=58.2
intel-openmp=2018.0.3
jpeg=9b
kiwisolver=1.0.1
libedit=3.1.20170329
libffi=3.2.1
libgcc-ng=8.2.0
libgfortran-ng=7.3.0
libpng=1.6.34
libstdcxx-ng=8.2.0
libtiff=4.0.9
libuuid=1.0.3
libxcb=1.13
libxml2=2.9.8
matplotlib=2.2.3
mkl=2018.0.3
mkl-service=1.1.2
mkl_fft=1.0.4
mkl_random=1.0.1
ncurses=6.1
ninja=1.8.2
nltk=3.3.0
numpy=1.15.0
numpy-base=1.15.0
olefile=0.45.1
openssl=1.0.2
pcre=8.42
pillow=5.2.0
pip=10.0.1
pycparser=2.18
pyparsing=2.2.0
pyqt=5.9.2
python=3.6.6
python-dateutil=2.7.3
pytorch=0.4.1
pytz=2018.5
qt=5.9.6
readline=7.0
setuptools=40.0.0
sip=4.19.8
six=1.11.0
sqlite=3.24.0
tk=8.6.7
torchvision=0.2.1
tornado=5.1
wheel=0.31.1
xz=5.2.4
zlib=1.2.11